feat: validatePositive(long, String) 추가 (In ValidationUtil) payEntity 충전 기능 추가 payTransactionRepository 추가 페이 거래내역 목적 및 타입 추가 (PayTransactionPurpose, PayTransactionType) BankAccountTransactionRepository 추가 BankAccountTransactionPurpose 내용 수정 BankAccountTransactionEntity 거래 금액 BigDecimal -> long 변경

+++ src/test/java/com/kt/domain/entity/BankAccountTransactionEntityTest.java
@@ -37,7 +37,7 @@ public class BankAccountTransactionEntityTest {
 				testUser.getBankAccount(),
 				BankAccountTransactionType.DEPOSIT,
 				BankAccountTransactionPurpose.SALARY,
-				BigDecimal.valueOf(10_000),
+				10_000,
 				BigDecimal.valueOf(10_000),
 				testUser.getBankAccount().getId()
 			);
@@ -52,7 +52,7 @@ public class BankAccountTransactionEntityTest {
 				testUser.getBankAccount(),
 				null,
 				BankAccountTransactionPurpose.SALARY,
-				BigDecimal.valueOf(10_000),
+				10_000,
 				BigDecimal.valueOf(10_000),
 				testUser.getBankAccount().getId()
 			)
@@ -71,7 +71,7 @@ public class BankAccountTransactionEntityTest {
 				testUser.getBankAccount(),
 				BankAccountTransactionType.DEPOSIT,
 				null,
-				BigDecimal.valueOf(10_000),
+				10_000,
 				BigDecimal.valueOf(10_000),
 				testUser.getBankAccount().getId()
 			)
@@ -90,7 +90,7 @@ public class BankAccountTransactionEntityTest {
 				testUser.getBankAccount(),
 				BankAccountTransactionType.DEPOSIT,
 				BankAccountTransactionPurpose.SALARY,
-				BigDecimal.valueOf(0),
+				0,
 				BigDecimal.valueOf(10_000),
 				testUser.getBankAccount().getId()
 			)
@@ -109,7 +109,7 @@ public class BankAccountTransactionEntityTest {
 				testUser.getBankAccount(),
 				BankAccountTransactionType.DEPOSIT,
 				BankAccountTransactionPurpose.SALARY,
-				BigDecimal.valueOf(10_000),
+				10_000,
 				BigDecimal.valueOf(-1),
 				testUser.getBankAccount().getId()
 			)
